+++ application.nix
@@ -83,18 +83,6 @@ rec {
                   ;;
               esac
 
-              # We want to avoid making the user configure audio outputs, but
-              # instead route audio to both the standard output and any connected
-              # displays. This looks for any "HDMI" device on ALSA card 0 and
-              # tries to add a sink for it. Both HDMI and DisplayPort connectors
-              # will count as "HDMI". We ignore failure from disconnected ports.
-              for dev_num in $(aplay -l | grep "^card 0:" | grep "HDMI" | grep "device [0-9]\+" | sed "s/.*device \([0-9]\+\):.*/\1/"); do
-                printf "Creating ALSA sink for device $dev_num: "
-                pactl load-module module-alsa-sink device="hw:0,$dev_num" sink_name="hdmi$dev_num" sink_properties="device.description='HDMI-$dev_num'" || true
-              done
-              pactl load-module module-combine-sink sink_name=combined
-              pactl set-default-sink combined
-
               # Enable Qt WebEngine Developer Tools (https://doc.qt.io/qt-5/qtwebengine-debugging.html)
               export QTWEBENGINE_REMOTE_DEBUGGING="127.0.0.1:3355"
 
@@ -140,7 +128,16 @@ rec {
 
       # Audio
       sound.enable = true;
-      hardware.pulseaudio = { enable = true; };
+      hardware.pulseaudio = {
+        enable = true;
+        extraConfig = ''
+          # Use HDMI output
+          set-card-profile 0 output:hdmi-stereo
+          # Respond to changes in connected outputs
+          load-module module-switch-on-port-available
+          load-module module-switch-on-connect
+        '';
+      };
 
       # Enable avahi for Senso discovery
       services.avahi.enable = true;