+++ controller/Changelog.md
@@ -7,6 +7,7 @@
 - os: Split system definition into base and application layers
 - os: Make build outputs and displayed system name configurable through application layer
 - os: Include live system ISO in deployed outputs
+- os: Update nixpkgs channel to 23.05
 - status screen: Display persistent storage usage statistics
 
 # [2023.2.0] - 2023-03-06